feat(mobile-app): Refectored by moving features package to base one
