+++ apps/mobile-app/src/main/kotlin/dev/marlonlom/apps/bookbar/features/books_new/NewBooksRoute.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package dev.marlonlom.apps.bookbar.ui.features.books_new
+package dev.marlonlom.apps.bookbar.features.books_new
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
