+++ apps/mobile-app/src/main/kotlin/dev/marlonlom/apps/bookbar/features/settings/UserSettingsViewModel.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package dev.marlonlom.apps.bookbar.ui.features.settings
+package dev.marlonlom.apps.bookbar.features.settings
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
