+++ apps/mobile-app/src/main/kotlin/dev/marlonlom/apps/bookbar/ui/main/scaffold/MainScaffold.kt
@@ -35,12 +35,12 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.currentBackStackEntryAsState
 import dev.marlonlom.apps.bookbar.core.preferences.UserPreferencesRepository
 import dev.marlonlom.apps.bookbar.domain.books.BookDetailResult
-import dev.marlonlom.apps.bookbar.ui.features.book_detail.BookDetailContent
-import dev.marlonlom.apps.bookbar.ui.features.book_detail.BookDetailsViewModel
-import dev.marlonlom.apps.bookbar.ui.features.books_favorite.FavoriteBooksUiState
-import dev.marlonlom.apps.bookbar.ui.features.books_new.NewBooksUiState
-import dev.marlonlom.apps.bookbar.ui.features.settings.UserSettingsDialog
-import dev.marlonlom.apps.bookbar.ui.features.settings.UserSettingsViewModel
+import dev.marlonlom.apps.bookbar.features.book_detail.BookDetailContent
+import dev.marlonlom.apps.bookbar.features.book_detail.BookDetailsViewModel
+import dev.marlonlom.apps.bookbar.features.books_favorite.FavoriteBooksUiState
+import dev.marlonlom.apps.bookbar.features.books_new.NewBooksUiState
+import dev.marlonlom.apps.bookbar.features.settings.UserSettingsDialog
+import dev.marlonlom.apps.bookbar.features.settings.UserSettingsViewModel
 import dev.marlonlom.apps.bookbar.ui.main.contents.AppContentCallbacks
 import dev.marlonlom.apps.bookbar.ui.main.contents.BookbarAppState
 import dev.marlonlom.apps.bookbar.ui.main.contents.rememberBookbarAppState
